GHA: Use julia-actions/cache action
See https://github.com/marketplace/actions/cache-julia-artifacts-packages-and-registry

This action caches artifacts and packages and hides the actual config similar to julia-runtest and julia-buildpkg. If a Julia update will require a change to the cache config, people using the template workflow won't have to manually go and fix it everywhere.

The action has options to disable caching of either packages and artifacts and to enable caching of registries, which may be worth considering as an option given to the user via the PkgTemplates interface.